+++ entrypoints/options/contexts/SettingsContext.tsx
@@ -45,6 +45,7 @@ const defaultSettings: AlfredSettings = {
   admin: {
     collapsibleSidebar: true,
     warnBeforeClosingCodeEditor: true,
+    themeListUtils: true,
   },
 };