+++ global.d.ts
@@ -69,6 +69,7 @@ declare interface AlfredSettings {
   admin?: {
     collapsibleSidebar?: boolean;
     warnBeforeClosingCodeEditor?: boolean;
+    themeListUtils?: boolean;
   };
 }