+++ wxt.config.ts
@@ -9,11 +9,18 @@ export default defineConfig({
     name: 'Alfred - Shopify Theme Detector',
     description:
       'Instantly detect themes on any Shopify store. Streamline your workflow with smart shortcuts and Shopify productivity tools.',
-    version: '2026.02.20',
+    version: '2026.02.23',
     action: {
       default_title: 'Alfred',
     },
-    permissions: ['contextMenus', 'scripting', 'tabs', 'activeTab', 'storage', 'webNavigation'],
+    permissions: [
+      'contextMenus',
+      'scripting',
+      'tabs',
+      'activeTab',
+      'storage',
+      'webNavigation',
+    ],
     host_permissions: ['<all_urls>'],
     web_accessible_resources: [
       {
@@ -42,7 +49,7 @@ export default defineConfig({
       // 'https://uptek.com',
       // 'https://junaid-workspace.myshopify.com/?pb=1&preview_theme_id=153167626454',
       // 'https://partners.shopify.com/1750954/stores/new?store_type=managed_store',
-      'https://junaid-workspace.myshopify.com/',
+      'https://admin.shopify.com/store/junaid-workspace/themes',
     ],
   },
 });